V2

+++ src-tauri/src/commands/calculator.rs
@@ -5,7 +5,7 @@
 #[tauri::command]
 pub fn calculate_price(
     material: f64,
-    tempo: i32,
+    tempo: f64,
     filamento_preco: f64,
     energia: f64,
     lucro: i64,
@@ -17,7 +17,7 @@ pub fn calculate_price(
 
     // Custo de energia: watts → kW × horas × preço/kWh
     let kwatts = watts as f64 / 1000.0;
-    let energy_cost = tempo as f64 * kwatts * energia;
+    let energy_cost = tempo * kwatts * energia;
 
     // Total com margem de lucro
     let total = (object_cost + energy_cost) * (1.0 + lucro as f64 / 100.0);
